cleanup code getting rid of several unused classes
